Tell users how to navigate the tour
It's not obvious how to navigate the tour. Upon first seeing it, I thought it was only one page, with the link at the bottom possibly providing that page in different languages. I think this one improvement will greatly increase the number of readers who make it past the first page.