Replaced some text...
I don't think it is wise to describe aspects of D functionality as "magic" since that gives the impression of something that people won't be able to understand and that is inconsistent/exceptional in relation to the rest of the language. So even if you don't accept my parenthasised note, I still hope you'll delete the "magic" text.